src/t2sql/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
python-dotenv
pydantic
sqlglot
openai
alembic
